requirements-dev.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements-dev.txt requirements-dev.in
#
alabaster==0.7.12         # via sphinx
appdirs==1.4.3
atomicwrites==1.3.0       # via pytest
attrs==19.1.0             # via pytest
babel==2.6.0              # via sphinx
beautifulsoup4==4.6.0
boto3==1.4.1
botocore==1.4.93
bravado-core==4.2.3rc7
collective.xmltestreport==1.3.2
colorama==0.4.1
crash==0.17.1
crate==0.16.1
croniter==0.3.4
docopt==0.4.0
docutils==0.14
elasticsearch==1.9.0
enum34==1.0.4
exifread==2.1.2
fake-factory==0.5.9
funcsigs==1.0.2           # via pytest
functools32==3.2.3.post2
future==0.17.1
futures==3.2.0
gevent==1.0.2
greenlet==0.4.15
httmock==1.2.5
idna==2.8
imagesize==1.1.0          # via sphinx
ipaddress==1.0.22         # via fake-factory
jinja2==2.8
jmespath==0.9.4
jsonpickle==0.9.3
jsonschema[format]==2.5.1
lovely-pytest-docker==0.0.5
lovely.esdb==0.3.8
lovely.essequence==0.1.0
lovely.pyrest==2.1.1
lovely.testlayers==0.6.3
mandrill==1.0.57
markupsafe==1.0
more-itertools==5.0.0     # via pytest
oauthlib==3.0.1
packaging==19.0           # via sphinx
pastedeploy==2.0.1
pathlib2==2.3.3           # via pytest
phonenumbers==8.0.0
pillow==5.4.1
pluggy==0.11.0            # via pytest
prompt-toolkit==1.0.16
py==1.8.0                 # via pytest
pygments==2.4.0
pyparsing==2.4.0          # via packaging
pyramid-jinja2==2.5
pyramid-swagger==2.3.1rc1
pyramid==1.7
pytest==4.5.0             # via lovely-pytest-docker
python-dateutil==2.4.2
python-magic==0.4.15
python-twitter==3.4.1
pytz==2019.1
pyyaml==3.11
repoze.lru==0.7
requests-file==1.4.3
requests-oauthlib==1.2.0
requests==2.7.0
rfc3987==1.3.1
s3transfer==0.1.13
scandir==1.10.0           # via pathlib2
short_url==1.2.2
simplejson==3.8.2
six==1.11.0
snowballstemmer==1.2.1    # via sphinx
sphinx-rtd-theme==0.1.6
sphinx==1.8.5
sphinxcontrib-plantuml==0.15
sphinxcontrib-websupport==1.1.2  # via sphinx
sphinxcontrib.httpdomain==1.7.0
strict-rfc3339==0.7
swagger-spec-validator==2.0.2
tldextract==2.0.1
transitions==0.4.1
translationstring==1.3
typing==3.6.6             # via sphinx
urllib3==1.25.2
validictory==1.1.2
venusian==1.2.0
waitress==1.3.0           # via webtest
wcwidth==0.1.7
webcolors==1.5
webob==1.8.5
websmscomtoolkit==1.0.2
webtest==2.0.18
z3c.recipe.scripts==1.0.1  # via collective.xmltestreport
zc.buildout==2.13.1       # via z3c.recipe.scripts, zc.recipe.egg
zc.recipe.egg==2.0.7      # via collective.xmltestreport, z3c.recipe.scripts
zope.deprecation==4.4.0
zope.exceptions==4.3      # via zope.testrunner
zope.interface==4.6.0
zope.testing==4.7         # via collective.xmltestreport
zope.testrunner==5.0      # via collective.xmltestreport
